refactor: outbound.ts -> outbound-types.ts 잘못된 파일명 수정
